patrimo_maintenance_module/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-dotenv
pandas
openpyxl
